After some small changes and failed attempt for segmentEdit
The EditSegments page is very bad and complicated to make work well and have a functional design at the same time. I'll give it up for now. If the person needs to edit the segment, it is best to delete the existing segment that is wrong and create a new one with the correct information. This way the user doesn't get confused and I don't have to do anything new.

Therefore, I just need to undo the routes and pages that I created for this editing function that I am abolishing